requirements.txt
streamlit
pandas
numpy
Pillow
matplotlib
scikit-learn
opencv-python
